requirements.txt
boto3==1.5.36
botocore==1.8.50
Django==2.0
django-storages==1.6.5
docutils==0.14
jmespath==0.9.3
mysqlclient==1.3.12
olefile==0.44
Pillow==4.3.0
python-dateutil==2.6.1
pytz==2017.3
s3transfer==0.1.13
six==1.11.0
